Move util stuff to new utils.py
